requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
decorator==5.1.1
Django==4.0.1
django-sslserver==0.22
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.6.1
emoji==2.12.1
exceptiongroup==1.2.2
executing==2.0.1
h11==0.14.0
idna==3.7
ipdb==0.13.13
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
numpy==1.26.4
outcome==1.3.0.post0
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
prettytable==3.10.2
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pyecharts==2.0.6
Pygments==2.18.0
pymongo==3.12.1
pyOpenSSL==24.2.1
PySocks==1.7.1
pytz==2024.1
requests==2.32.3
selenium==4.2.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.2.4
stack-data==0.6.3
tomli==2.0.1
tqdm==4.66.4
traitlets==5.14.3
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==1.26.19
urllib3-secure-extra==0.1.0
wcwidth==0.2.13
wsproto==1.2.0
